LayoutTests/fast/page-color-sampling/color-sampling-ignores-empty-container.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=50 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body {
            font-family: system-ui;
            height: 200vh;
        }

        footer {
            position: fixed;
            bottom: 0;
            left: 0;
            width: 100%;
            height: 100px;
            color: white;
            text-align: center;
            z-index: 1000;
        }

        button {
            background: rgb(52, 199, 89);
            width: 80%;
            height: 40px;
            border-radius: 4px;
            border: none;
            font-size: 16px;
            color: white;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(50, 0, 50, 0);
        await UIHelper.ensurePresentationUpdate();
        colors = await UIHelper.fixedContainerEdgeColors();
        shouldBeNull("colors.top");
        shouldBeNull("colors.left");
        shouldBeNull("colors.right");
        shouldBeNull("colors.bottom");

        finishJSTest();
    });
    </script>
</head>
<body>
<div class="tall"></div>
<pre id="console"></pre>
<pre id="description"></pre>
<footer><button>Sign up for free</button></footer>
</body>
</html>
